ref(deletes): bulk delete consumer (#6510)
**context**:
The work for bulk deleting in snuba has so far included the following:
* Added the kafka schemas -
https://github.com/getsentry/sentry-kafka-schemas/pull/347
* Adding the new endpoint including producing to the topic -
https://github.com/getsentry/snuba/pull/6440
* Created the topic in production for all env -
https://github.com/getsentry/ops/pull/12711

**what's left**:
Now that we have the topics created we can finish up the consumer side.
- [ ] Add the consumer logic to snuba (This PR)
- [ ] Add the consumer deployment to S4S region in the ops repository
- [ ] Set up datadog alerts/metrics and other observability

**this PR**:
It's a bit of a larger PR but it can be reviewed in a couple sections:
* The main consumer logic and strategy
* This has the logic to create the strategy factory for the consumer and
composes the strategy steps. The `strategy.py` file has the details for
actually executing the delete query
* The `batching.py` file - I have an [arroyo
PR](https://github.com/getsentry/arroyo/pull/390) that makes this file
obsolete but in the meantime I don't think it needs to block this PR
since it will be easy to remove after
* The formatters are going to be the only code that someone will need to
write in the future when deploying the deletions consumer for a
different storage. How one formats the conditions for the `DELETE` query
is up to that logic.