feat(profiler): Update formatting of Heaptrack output file (#6601)
Since Heaptrack performs continuous profiling of the Rust consumer, we
will want to track the profiler's output for various profiling runs that
happen. A profiling run happens for every consumer process that is
started.

Adds timestamp and ensures Heaptrack writes to a formatted file for each
profiling run. This will also make analysis easier as we will know which
file corresponds to which date and time.