feat(runtime_config): A way to add static descriptions to config items
Ability to document runtime config in-code. This should be used to
describe what the option does, while the existing description field may
be used to document why the option was set this way.

Also get rid of some tech-debt. We were pulling in random styles in the
capacity management project. This caused the navbar to change styles
when switching to that tab. The other style changes in nav.tsx are there
to deal with the now-imported bootstrap.css correctly.