feat(runtime_config): A way to add static descriptions to config items (#6624)
Ability to document runtime config in-code. This should be used to
describe what the option does, while the existing description field may
be used to document why the option was set this way.

Also get rid of some tech-debt. We were pulling in bootstrap via
jsdelivr
in the capacity management project. This caused the navbar to change
styles
when switching to that tab. The other style changes in nav.tsx are there
to deal with the now-imported bootstrap.css correctly.


![image](https://github.com/user-attachments/assets/1613c8a2-a3cc-45a1-8a69-a5fddcd5fad2)