consumer: Add flag to configure DLQ buffer limit (#6626)
Arroyo consumers maintain an in-memory buffer of messages that gets
filled every time a new message is polled from rdkafka. Offsets get
removed from the buffer when messages are committed.

The buffer is kept around in order to populate the DLQ topic and replay
invalid messages later.

However, this buffer can grow unbounded since we haven't configured any
cap on it. We have a hunch that this is also contributing to memory
spikes on Arroyo Rust consumers.

Add a flag to configure a limit. We seem to already [have the
functionality
](https://github.com/getsentry/arroyo/blob/main/rust-arroyo/src/processing/dlq.rs#L413-L421)to
actually set it in the DLQ buffer.