feat(release): Replace release bot with GH app (#6646)
The `getsentry-release` is a GitHub bot account that is used in various
automation, and the `${{ secrets.GH_RELEASE_PAT }}` is a personal access
token from that bot account. We are using a regular GitHub account as a
bot while it should be a non-human account since there are no humans
behind it. Hence, we are replacing it with a GitHub App.

Functionality wise, there will be no difference. 

More details:
https://www.notion.so/sentry/DACI-Replace-GitHub-bot-accounts-with-GitHub-Apps-getsentry-release-15109965d1204a91b9be71c49e8b66e0?pvs=4

#skip-changelog