ref(lw-deletes): add project_id killswitch and some logging (#6677)
This PR has a couple improvements and additions to the lw delete
pipeline, the main one being the killswitch for the project.

Most storages should have the `project_id` as one of the allowed columns
so the assumption right now is that they do and so the killswitch will
work for that case. If we add more storages in the future that don't
have that, we can update the killswitch logic. The config will be based
on the storage:

`lw_deletes_killswitch_search_issues`

The killswitch is added at the API level because the moment it's
produced to the consumer it's batched with other messages, most likely
with other `project_id`'s. Another option would be to add another
strategy step prior to the batch step that filters based on the
killswitch, but I felt like not producing to the topic at all made more
sense right now.