chore(inc-984): Add basic PII detector
In INC-984 we had an incident where IP addresses were copied into a new
unscrubbed field. It took us too long to detect that issue.

This metric could be used in combination with anomaly detection to
detect such faulty rollouts more quickly, but not prevent them.

TBD is who actually gets those alerts and how we should react to them.