fix(eap): fix bug with data_present (#6726)
Fixes https://github.com/getsentry/eap-planning/issues/144

## Additional Context
When we perform aggregations over attributes, the function is converted
into function_nameIf (e.g. countIf) which returns a default value if no
rows match the condition. This means that there is no way to distinguish
between an aggregate result that is 0 because it's for example a sum of
values that add up to 0, and a result that is 0 because no values
matched the given condition. To deal with this, we compute the number of
events being aggregated even when we aren't extrapolating so we can
determine if data was present or not.