feat(devservices): Support no workers option for containerized version of snuba (#6737)
This option should be supported as CI in sentry generally runs snuba
devserver without workers. Currently, this is determined in sentry based
on the `SENTRY_EVENTSTREAM` setting. This is a workaround for an issue
that stems from our jumbled logic around configuration files and
settings that we should discuss in the future.