chore(deps): Bump Python to 3.11.11 (#6719)
There are a number of security vulnerabilities which are present in the
Python 3.11.8 image, so bumped to the latest version of 3.11 which is
3.11.11.

Fixes #6718

### Legal Boilerplate

Look, I get it. The entity doing business as "Sentry" was incorporated
in the State of Delaware in 2015 as Functional Software, Inc. and is
gonna need some rights from me in order to utilize my contributions in
this here PR. So here's the deal: I retain all rights, title and
interest in and to my contributions, and by keeping this boilerplate
intact I confirm that Sentry can use, modify, copy, and redistribute my
contributions, under Sentry's choice of terms.

Co-authored-by: Markus Unterwaditzer <markus-honeypot@unterwaditzer.net>