fix(settings): specify VALID_RETENTION_DAYS for self-hosted (#6756)
Fixes https://github.com/getsentry/self-hosted/issues/3421

I saw from the rust_snuba code, that we have this:
https://github.com/getsentry/snuba/blob/acfbe9a31880ecfe15888386bf40a2afe2b16088/rust_snuba/src/processors/utils.rs#L10-L20

But if we follow the default value, it will always defaults to `30`.
Whereas not everyone on the self-hosted instance uses value greater than
30. My own instance is set to 14.

https://github.com/getsentry/snuba/blob/acfbe9a31880ecfe15888386bf40a2afe2b16088/snuba/settings/__init__.py#L251

### Legal Boilerplate

Look, I get it. The entity doing business as "Sentry" was incorporated
in the State of Delaware in 2015 as Functional Software, Inc. and is
gonna need some rights from me in order to utilize my contributions in
this here PR. So here's the deal: I retain all rights, title and
interest in and to my contributions, and by keeping this boilerplate
intact I confirm that Sentry can use, modify, copy, and redistribute my
contributions, under Sentry's choice of terms.