feat(eap): allow turning off subscriptable mapping in query settings (#6785)
This was introduced to the query pipeline stage before there was a
clickhouse map type. Now certain datasets in EAP do not use it. Allow it
to be turned off via query pipeline.


I thought about doing this via a storage setting but it felt like more
work than necessary