fix: make snuba admin error messages more stable (#6778)
this PR makes the error handling in snuba admin more robust to hopefully
prevent errors messages from getting messed up in the UI as frequently.

it maintains the console error logging as backup.

## Major changes
* adds `_convertAnyToError` function that is responsible for converting
any type object to an error object. This is the change that has most
impact on the already existing messages. It replaces the logic that was
in `defaultError` but is more robust.
* changes the type of `onError` property of execute button to
`function(Error)` instead of `function(Any)`. this makes overriding the
error handling behavior simpler to users of the button. I was the only
user of this so it has no serious impacts.