feat(admin): Add a standardized component for copying query results
At lot of components offer the ability to copy the results of a run query, but each had a slightly
different way of providing that functionality. That meant that not all functionality across the
components.

Add a standard component for these buttons. This will also ensure that system queries now have a
CSV copy results.