feat(outcomes): Add buffer_overflow discard reason (#6792)
ref https://github.com/getsentry/team-sdks/issues/116

This PR implements a new client discard reason for `buffer_overflow`.
This will be used to track when the internal breadcrumbs buffer
overflows for the new logs product that we are working on. This is
documented in develop here:
https://github.com/getsentry/sentry-docs/pull/12395

I also went ahead and added code comments that describes what all the
different client discard reasons mean. This is done in commit
b3dcd678446b89109dac6ef9ca28c8fd929073d0, while the actual change to add
the `buffer_overflow` discard reason was added in
dd25dde8926b16838f2ba2bdb9f1de06a8a4ec30.

Note: The reason we have `buffer_overflow` as a separate item to
`queue_overflow` is that in the future when we send log items in
envelopes we'll increment `queue_overflow` for the transport queue. We
want to differentiate between the transport queue and the internal
buffer explicitly.

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>