fix(eap-api): Snuba receives and returns doubles, while still supporting floats (#6782)
Update plan:

1. Sentry can receive doubles and floats
2. Snuba can receive (from Sentry) and return (to Sentry) doubles and
floats. In the case of aggregation, replace float with double. This is
fine because Sentry handles the case of receiving doubles already
3. Sentry asks for doubles only

This PR addresses step 2. Step 1's PR is here:
https://github.com/getsentry/sentry/pull/83566

For every test that involved floats, I added a corresponding test with
doubles. It ensures backward compatibility by verifying that Snuba can
handle both floats and doubles

---------

Co-authored-by: Rachel Chen <rachelchen@PL6VFX9HP4.attlocal.net>
Co-authored-by: Rachel Chen <rachelchen@PL6VFX9HP4.local>