dev(rate-limits): Increase backfill query allocation to 60 per s (#6338)
Based on the Datadog metrics[1] and the corresponding Sentry error[2], the `getsentry.tasks.backfill_grouping_records` referrer is hitting our 30/s rate limit pretty frequently.

We would like to increase to 60/s if okay with team. The queries that we're making are point queries, so they should not scan very much data. 


[1] https://app.datadoghq.com/dashboard/tht-dpn-5ui/allocation-policy?fromUser=false&fullscreen_end_ts=1727203031641&fullscreen_paused=false&fullscreen_refresh_mode=sliding&fullscreen_section=overview&fullscreen_start_ts=1726598231641&fullscreen_widget=984243635966760&refresh_mode=sliding&view=spans&from_ts=1727169007884&to_ts=1727183407884&live=true
[2] https://sentry.sentry.io/issues/5655838252/