hotfix(ci): Fix missing image sha output (#6813)
The output of the build job has not been properly declared. ~Instead of
getting it to work, let's just hardcode the image, it's actually
simpler.~ It turns out that the e2e-test job does not have access to the
right github.sha for some reason, probably because it doesn't check out
the snuba repo?