fix(admin): Make clickhouse node info more retriable (#6808)
Sometimes in snuba admin we fail to load the cluster info because one
cluster rejected a connection, as we add more clusters, the probability
of this will only increase. Add a retry to the function. WIll add
caching of this info later as it's not needed to be fresh fetched all
the time