feat: eap EndpointTraceItemTable spare aggregate attribute transformer (#6817)
this pr addresses this ticket for EndpointTraceItemTable.
https://github.com/getsentry/eap-planning/issues/158

It makes it so when we have a query like
```
SELECT sum(wing.count)
...
GROUP BY animal_type
```
it will add
```
SELECT sum(wing.count)
WHERE hasAttribute(wing.count)
...
GROUP BY animal_type
```

## Major change
The major change in this PR is the addition of
`SparseAggregateAttributeTransformer` this is a class responsible for
transforming the original `TraceItemTableRequest` into the new one with
the additional filter. This query transformation step is then added as a
step in the resolver that all requests go through.


I think it should be sufficient to only do EndpointTraceItemTable and
not EndpointTimeSeries since timeseries wont graph the 0 data but ill
ask tony to make sure. If timeseries is needed it should only be 1-2
days of extra work.

---------

Co-authored-by: Onkar Deshpande <onkar.deshpande@sentry.io>