feat: Add EndpointTraceItemStats RPC (#6809)
requires https://github.com/getsentry/sentry-protos/pull/101, will bump
package once it's merged

For comparison workflows, we need the count distribution data of all
attribute/value pairs (both string and numeric, however we'll focus on
string first). We're introducing a trace item stats endpoint to
facilitate that. The aim is to build a somewhat generic endpoint that
can give a high level overview of the shape of your data.

This implementation is fairly simple, restricted to string attributes to
begin with. It returns the frequency distribution of attributes in the
requested filter. The data returned is ordered by keys that have the
highest frequency attribute values.

This first pass aims to also unblock work the ML team is doing to
finalize the algorithm to rank these attributes.