fix(eap): Remove static sample count thresholds (#6892)
The `reliability` flag is supposed to return the mathematical accuracy
based on the sample rate without other additional heuristics applied.
Currently, the calculation also compares the sample count with a static
threshold (count > 100) and returns _low reliability_ if the count is
below.

Going forward, any checks for a static sample count threshold will be
performed outside of the EAP.

Confidence calculation for aggregations other than AVG will be fixed in
follow-up PRs.