fix(eap): Update accuracy formulas for sum, count, and avg (#6923)
Updates the formulas to properly implement the semantics of the
reliability (accuracy) flag:
- Drops the logarithm for count. This was added as a heuristic to make
`counts` more sensitive to low sample rates, but that is not
mathematically established and ended up flagging too many count queries.
- Similarly, apply the same kind of variance computation for sum.
Particularly, this ensures that accuracy is always `True` when the
sample rate is at 100%, which is the expected outcome. The previous
formula did not behave like this and ended up flagging in accounts with
100% sampling.
- For average, we combine the variance of sum and count using the
Bonferroni method. However, since we're only interested in the relative
error, we take the average of the upper and lower bound deviation.

For more information on the underlying math, see this [Cloudflare
Blogpost](https://blog.cloudflare.com/how-we-make-sense-of-too-much-data/#extracting-value-from-downsampled-data)

The formulas are thus:

- **Count**: $Z \cdot \sqrt{\sum_{i=1}^n w_i^2 - w_i}$
- **Sum**: $Z \cdot \sqrt{\sum_{i=1}^n x_i^2 \cdot (w_i^2 - w_i)}$
- **Avg**: $(\\frac{t + err_t}{c - err_c} - \\frac{t - err_t}{c +
err_c}) \cdot 0.5$

See code comments for more description.