feat: str_attr and num_attr materialized views for new eap_items table (#6907)
the purpose of this PR is to recreate the existing
`spans_num_attrs_3_mv` and `spans_str_attrs_3_mv` so they read from
`eap_items_1_local` (they used to read from `eap_spans_2_local`). it
addresses this ticket
https://github.com/getsentry/eap-planning/issues/194.

## major changes
* create a materialized views `items_attrs_1` via migration
* creates a storage definitions for it
* no more aggregations in the mv, there used to be `count`, `min_value`,
and `max_value` but we never used them

## design decisions
* attr_type indicates whether the attribute is a float or string, I made
this a low cardinality string with values `string` and `float`
* originally i made attr_value nullable bc its not present for
`attr_type='float'` but this made it so i cant use attr_value in the
order by. so now attr_value just gets set to empty string for floats.
* I chose ReplacingMergeTree for the MV to save space by deleting
duplicate rows