fix(eap): Update relative confidence calculation (#6935)
The computation of the reliability threshold did not work correctly when
the column value is negative. Given a `value` and an absolute error
`error`, the computation was: `(value + error) / value <= 1.5`. If the
value is positive, this computation is correct: It increases the
numerator resulting in a ratio >= 1.0. Since `error` is always positive,
for negative values this however results in a ratio <= 1, which will
always report high confidence.

To fix this, we now compare the error directly with the value and change
the threshold constant from 150% (upper bound) to 50% (relative error).
The new formula therefore is `abs(error/value) <= 0.5`. The same fix
applies to the percentile extrapolation context, where we first compute
the difference between the upper bound and value as error, and then
compare in the same way.