feat(eap-spans): push eap-items commits to the commit log (#6990)
We want this in order to run eap_spans subscriptions from the eap_items
commit log without having to switch subscription scheduler and executor
topics or run a parallel processing pipeline. These are fed from the
same topic (snuba-spans) and so the commit offsets/timestamps should be
comparable.

This change starts pushing eap_items commits to the spans commit log
topic. The scheduler consumer has logic to ignore these messages until
an ops configuration change is made. We add a metric here to make sure
that path is hit