feat(ci): tag GHCR image with release version (#6997)
Similar work with these PRs:
- https://github.com/getsentry/relay/pull/4532
- https://github.com/getsentry/symbolicator/pull/1635

While also trying to provide a solution (or at least an alternative) for
this issue: https://github.com/getsentry/self-hosted/issues/3593


### Legal Boilerplate

Look, I get it. The entity doing business as "Sentry" was incorporated
in the State of Delaware in 2015 as Functional Software, Inc. and is
gonna need some rights from me in order to utilize my contributions in
this here PR. So here's the deal: I retain all rights, title and
interest in and to my contributions, and by keeping this boilerplate
intact I confirm that Sentry can use, modify, copy, and redistribute my
contributions, under Sentry's choice of terms.