Fix: Add correct clickhouse connection http scheme in rust consumer (#7034)
I encountered the need to use Snuba with a secure connection to
ClickHouse, but it turned out that the URL for the query hardcodes the
http scheme. So I added a parameter for a secure connection in the
connection configuration to the Rust consumer and implemented logic to
determine the required scheme (http/https) based on it.

<!--

  Sentry employees and contractors can delete or ignore the following.

-->

### Legal Boilerplate

Look, I get it. The entity doing business as "Sentry" was incorporated
in the State of Delaware in 2015 as Functional Software, Inc. and is
gonna need some rights from me in order to utilize my contributions in
this here PR. So here's the deal: I retain all rights, title and
interest in and to my contributions, and by keeping this boilerplate
intact I confirm that Sentry can use, modify, copy, and redistribute my
contributions, under Sentry's choice of terms.

---------

Co-authored-by: Markus Unterwaditzer <markus-honeypot@unterwaditzer.net>