feat: eap timeseries works for logs (#7047)
This PR allows timeseries to work for logs. 

It take the timeseries resolver for eap_spans and extracts it into a
generic eap_items resolver. Now eap_spans and eap_logs both use the
eap_items resolver. This also enables eap_logs to work with timeseries
now (it used to not).

## major changes
* moved eap_spans resolver into new eap_items one
* changed eap_spans and eap_logs resolver to call the eap_items one
* added a test to ensure timeseries works for logs