feat(sampling-in-storage): most downsampled tier is now tier 64 (#7067)
Based on manual QA, we see tier512 being a small amount of data for
small orgs which doesn't provide a good way to extrapolate (hence why we
also bump the threshold).

On top of that, we noticed the data we show until we display the result
of the higher tier query is meaningfully different.

Based on this, bumping the minimal tier would make us query more data in
hopes of better accuracy when estimating what higher tier to hit and
also present less of a difference in the shape of the data.

https://github.com/getsentry/snuba/pull/7066#issuecomment-2798106164

---------

Co-authored-by: Rachel Chen <rachelchen@MacBookPro.attlocal.net>
Co-authored-by: Rachel Chen <rachelchen@PL6VFX9HP4.local>