feat(sampling_in_storage): Log routing decisions to querylog (#7096)
Right now we don't have a good way to drill down and understand our
routing decisions. spans do an okay job but are subject to sampling.
This PR:

* creates a hacky entry in the querylog for every routing decision we
make, it abuses the `clickhouse_queries.stats` field which accepts
arbitrary JSON and the `dataset=storage_routing` field to make the
requests separate
* All metrics outputted by the routing strategy are added to the payload
* Adds tests for this functionality