feat(errors) - Populate new `timestamp_ms` on errors (#7123)
We are now parsing the `datetime` param arriving on messages into the new `timestamp_ms`,
preserving its microsecond part. There was actually an issue with the parsing there which this commit corrects. 
This parsing did not matter until now because the sub second part was always ignored as we saved
into Datetime fields (which have a 1 second resolution by default) on both Errors and Outcomes.