feat(storage_routing): Use tiers lower than TIER_8 after 10 billion ingested items (#7141)
After turning on the outcomes based strategy for all customers, we're
seeing an exceeding of the time limit more than we would like

https://app.datadoghq.com/s/FH6-Y3/889-thi-r3u

spot checking a few, it seems like the ingested items were in the tens
of billions for those queries.

Add the ability to downsample further. This can be tuned by changing the
`OutcomesBasedRoutingStrategy.max_items_before_downsampling`
configuration in the runtime config