ref(eap): remove timeout_overflow_mode (#7146)
**context**
The
[`timeout_overflow_mode`](https://clickhouse.com/docs/operations/settings/settings#timeout_overflow_mode)
setting in ClickHouse is suppose to dictate the behavior of a query if
it exceeds `max_execution_time`

options for `timeout_overflow_mode`:
* `throw` which throws a `TimeoutExceeded` error (the default)
* `break` which is supposed to return partial results

Setting `timeout_overflow_mode` to break doesn't actually provide the
partial results that it was assumed to return. Additionally, not only
does it _not_ provide partial results, but it also is likely related to
the series of `TablesStatusResponse` errors we have been seeing when
testing out the newer version of ClickHouse that we want to upgrade to.

In order to prevent possible issues when doing the rest of the
ClickHouse upgrades - I'm removing any places where we set this to
`break`

**product considerations**
Right now since we aren't getting proper results back, I don't think
this should change how we are rendering the results. I think we get an
error anyway, it's just not a timeout error (this doesn't happen super
ofter but here is an example from snuba admin us)
<img width="1336" alt="Screenshot 2025-04-29 at 12 30 30 PM"
src="https://github.com/user-attachments/assets/294bfa4a-a1b2-42f8-ae7a-e39d4d08b608"
/>