fix(trace-items): Should return keys in filter (#7162)
When suggesting keys, we should still return the keys that are already
in the filter. This is because as a user, we'd still expect that key to
show up in the autocompletions. One case where this is best observed is
if the filter is on a wild card and they want to narrow the search.
Removing the key from the suggestions seems premature and can lead to
confusion.