fix: fix duplicate labels in RPC causing incorrect behavior (#7139)
The main purpose of this PR was to fix a bug involving duplicate labels
provided in a users request. Top level expressions have labels,
additionally the nested aggregates and formulas inside the expressions
also have labels. EX:
```
expression: {
  label: label1
  aggregate: {
    label: label2
    ...
  }
}
```
When the inner labels are duplicated across expressions it causes
incorrect behavior. To see the specific bug that was fixed, see here:
https://github.com/getsentry/eap-planning/issues/242

this PR resolves https://github.com/getsentry/eap-planning/issues/244

## Major Changes
I solved this by implemented visitors on the protobuf request do the
following:
1. ensure all top-level expressions have labels, if not adds a unique
one. this is required for later steps.
2. remove all inner-expression labels. that is all labels inside an
expression that isnt the very top level label. This solves the bug of
duplicate labels.
3. add a label to all `aggregate` or `conditional aggregate` expressions
that is the same as the expression label. I discovered while making this
PR that this is necessary for correct functioning of the timeseries
endpoint.

4. In order to create all these visitors, I implemented a new base
visitor class that I believe offers some benefits over our existing one.
The main benefit is that this one each visitor can independently define
traversal, whereas with the old one all visitors had to traverse the
same.

## Next steps
This PR makes any user provided labels inside an expression obsolete, so
those fields should be deprecated in the protobuf.