chore(ci): Bump sentry-related dependencies more aggressively (#6769)
Open PRs to update sentry-sdk, sentry-arroyo at once, and with at most 1
day delay. It should not actually create PRs every day because it's only
two dependencies that get released rarely.

context: we were behind a few arroyo versions. when we had to bump the
arroyo version during incident response for INC-1003, we got held up on
breaking changes that had been released a few days ago but not been
bumped in snuba.

sentry-kafka-schemas is not there because its release speed is way too
high.