feat(consumers): Create an experimental healthcheck using a different strategy (#7154)
Instead of touching the health file every time we call poll, we only
touch it when one of the following two conditions is true:
1. We have a commit request which means we are actively processing
messages
2. We are calling poll after not calling submit in the previous
iteration - this happens when we don't have messages to process which is
a healthy state

Here's an example of how we could enter an unhealthy state:
- If we are not getting back a commit request in poll, it means there's
either a strategy error (an unhealthy state) or we haven't committed and
thus not making progress (potentially unhealthy if we are still
receiving messages). This allows us to kill the consumer using the
liveness probe (i.e. if we are in an unhealthy state for a certain
amount of time, the consumer will be killed)