feat(devservices): Add containerized modes for profiles, uptime, metrics dev (#7229)
This adds more containerized modes that start up groups of consumers to
support profiles, uptime, and metrics dev. The goal of this is to allow
toggling of features in sentry without the injection of env variables

Context: 
We're trying to ease development within Sentry if devs want to enable
features such as profiling, uptime, and metrics. Instead of changing
django settings in their sentry config and using env variable injection
manually to start various groups of consumers to support these features,
we'd like to make it easier.

This allows us to support something like devservices up --mode profiling
in sentry where users would have all the ingest consumers for profiling
brought up, vroom, and snuba consumers within one command. Devservices
is able to import the definitions to understand the groups of snuba
consumers that need to be brought up to support certain features.