feat(consumers): Re-implement batching consumer (#7248)
This is effectively a revert of
https://github.com/getsentry/snuba/pull/5694

### Reasoning

The streaming writes consumer saves memory however, by attaching
together the batching and writing step, it makes it difficult to
implement the following features:

1. A retry policy for failed clickhouse writes
2. Dynamic batch sizes in response to clickhouse errors

Since the original implemenation, Arroyo has changed and so has the
consumer codebase, this PR is bringing those two worlds together

### Testing

This was tested by running the items consumer locally and running the
script in this PR to generate messages. It was verified that messages
did make it over to clickhouse

### Rollout

As of writing, the v2 consumer only runs on eap consumers in S4S. We
need to monitor the impact that this change has to memory usage and
consumer performance before we go further with rollout or optimizations.