fix(TraceItemsDetails): dedupe int,bool and float attrs (#7260)
Addresses
[EAP-121](https://linear.app/getsentry/issue/EAP-121/traceitemdetails-endpoint-should-merge-attributes-without-leaving)

For `bool` and `int` attributes, we also store a `float` version in the
database to simplify aggregations. When returning data to the user we
should only return the original type, and ignore the float duplicate.

Waiting on https://github.com/getsentry/sentry/pull/94553 to be merged
before I can merge this PR