ref(subscriptions): raise SubscriptionQueryException on query exception (#7299)
This is a followup to INC-1238 where we didn't catch subscriptions query
errors in canary before going to prod. Example query exception that was
missed
[SNUBA-954](https://sentry.sentry.io/issues/6734654425/?environment=prod&environment=us&environment=production&project=300688&query=server_name%3Asnuba-eap-spans-subscriptions-executor%2A&referrer=issue-stream&stream_index=4)

So when digging into this, I don't think we were actually using the
try/catch that was added in `submit`. The `poll` method is where we
actually get the results of the future by calling
`result_future.future.result()` that is when we can handle the
exception, according to the
[docs](https://docs.python.org/3/library/concurrent.futures.html#concurrent.futures.Future.result).

So I've moved the try/catch to the `poll` method and then added the
functionality to raise a `SubscriptionQueryException` so that it's
explicit when a query exception in the subscriptions versus the regular
query pipeline