feat(eap): add BinaryFormula to AggregationComparisonFilter (#7314)
Allow for BinaryFormula style filtering based on the results of
aggregations. This doesn't buy us much in terms of performance (all the
aggregations still need to be computed) but it's convenient for callers.

Requires https://github.com/getsentry/sentry-protos/pull/133, tested
with:
```
$ pip install -e ../sentry-protos/py --config-settings editable_mode=strict
```

The test failing in CI is passing

[[internal
ticket]](https://linear.app/getsentry/issue/EAP-74/allow-aggregationcomparisonfilter-to-support-binaryformula)