feat(errors): Add group_first_seen column to issues platform storage (#7336)
We want to denormalize the `first_seen` column from the
`grouped_messages` table to the `search_issues` table.

The purpose here is that we want to be able to sort queries on `errors`
by the group's first-seen time.

This sounds similar to #7305 because it is... we discovered that the
issues search uses data from both tables, so both will need the new
column.