fix(storage-routing): If item type is unspecified, default to spans (#7350)
Requests for Traces do not have item type specified (which makes sense).
In order to use tracing, a customer is most likely going to be ingesting
spans. Thus, use spans as the default item when unspecified.

We can definitely get more sophisticated with this but this will allow
us to sample GetTraces calls and solve some timeouts on big customers