fix(issues): Start processing new_group_first_seen in merge case (#7351)
This is part of an effort to fix the issue-search's sort-by-age feature.
As part of that feature, we are adding the "first_seen" time of a group
to each error. When we merge groups, we want to update the errors to
have the new group's "group_first_seen".

We've decided to do that by explicitly passing the primary group's
"first_seen" time along with the Kafka message. This PR starts ingesting
that field properly; a followup PR in the `sentry` repo will start
providing that field after this has deployed everywhere.