feat: support reliability for formulas in timeseries endpoint (#7331)
This PR enables reliabilities to be returned for formulas. The
reliability of a formula is defined to be reliable iff each of its parts
is reliable. ex: `(agg1 + agg2) / agg3` is reliable iff `agg1, agg2,
agg3` are all reliable.

## implementation
its implemented in 2 main parts
1. step 1, add the leaf nodes of a formula to the request. when a user
requests `(agg1 + agg2) / agg3` we traverse this formula to also add
`agg1, agg2, agg3` to the request. this is necessary to know their
reliabilities are later on. they will have the labels set to
`form_label.left.left, form_label.left.right, form_label.right`
respectively.
2. step 2, traverse the results to find the reliabilities of the leaf
nodes and use them to compute the reliability of the formula. and then
afterwards remove the extra colums `agg1,agg2,agg3` from the result so
we dont actually return it to the user when they didnt explicitly
request it (remember these were added in step 1).

# testing
I added a single large unit test that tests reliability of formula works
in 3 different cases.