feat(search_issues): add `timestamp_ms` column on `search_issues` (#7358)
Similar to https://github.com/getsentry/snuba/pull/7115 (in effect) and
https://github.com/getsentry/snuba/pull/73345 (in implementation). We
want more precision on replays breadcrumb ordering, and replays query
both errors and issue platform, which is `search_issues` in snuba, so we
want to add an additional column that has millisecond precision. Column
is exactly the same as `timestamp_ms` in errors.

Refs
https://linear.app/getsentry/issue/ID-605/precise-error-ordering-on-replay-breadcrumbs-queries