ref(outcomes): add daily local and mv tables (#7356)
The outcomes table is used as part of our routing strategy to determine
when to query downsampled data. Now that we are making the move to
increase the retention of our data, we'll want to be able to query more
than 90 days worth of outcomes to route data accordingly.

The outcomes hourly table in US has about 2GB worth of data per
partition (it's actually a bit less), and we partition by week. So
assuming we increased the TTL on that table to 13 months, thats 56
weeks.

```
56 weeks * 2GB = 112GB
```

And with the daily table that should cut down even more since we'll be
rolling up by day, not hour

```
112GB / 24 = ~5GB
```

Even if these numbers are slightly off, this shouldn't add a ton of load
to the outcomes cluster. And the outcomes consumers should have more
than enough headroom to accommodate the overhead a matview adds to
inserts