fix(search_issues): stop truncating timestamp_ms in search_issues (#7373)
In https://github.com/getsentry/snuba/pull/7362, we populate
`timestamp_ms` in `search_issues` by just copying the `timestamp` of the
event, since the event timestamp had all the precision we needed. I had
assumed that Clickhouse was the one that truncated the `timestamp` to
seconds, but it turns out there's also JSONRowEncoder, which forcefully
takes `datetime` objects and chops off the milliseconds:
https://github.com/getsentry/snuba/blob/799e027def8bc187c3eb4353b635baee908060fc/snuba/clickhouse/http.py#L54-L55


https://github.com/getsentry/snuba/blob/7bfa538516882dc58b2bed2f3e93e87360a6719c/snuba/clickhouse/__init__.py#L1

I don't want to change DATETIME_FORMAT, since that would be very
impactful to a lot of existing schemas. To work around this, we make
`timestamp_ms` an actual timestamp in milliseconds and pass it to
clickhouse, which then successfully does the conversion to
DateTime64(3), and will populate `timestamp_ms` with timestamps that
have the appropriate amount of precision.

One alternative I considered was wrapping `timestamp_ms` in a special
DateTimeWithMorePrecision class and special-casing it in `http.py`, but
I think that just adds more indirection for this specific feature — if
we have more columns that need additional precision that are managed by
JSONRowEncoder, I think that would be an appropriate followup.