fix(observability): raise a specific error for query timeouts (#7364)
We have sentry errors clogging our feed whenever requests time out. Most
of these are requests from HIGHEST_ACCURACY queries, which we expect to
time out some of the time.

Raise a different error for timeouts based on the downsampling mode so
we can know how many normal requests are timing out and we should take
action on

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>